2212 - Revised the attachment factory API
